Moving merge simulator, and applying it to the const_write_amplification_merge_policy
