+++ quickwit/quickwit-indexing/src/actors/merge_planner.rs
@@ -155,237 +155,3 @@ fn belongs_to_pipeline(pipeline_id: &IndexingPipelineId, split: &SplitMetadata) 
         && pipeline_id.node_id == split.node_id
         && pipeline_id.pipeline_ord == split.pipeline_ord
 }
-
-#[cfg(test)]
-mod tests {
-    use std::collections::{BTreeSet, HashMap};
-    use std::iter::FromIterator;
-    use std::ops::RangeInclusive;
-
-    use proptest::sample::select;
-    use quickwit_actors::{create_test_mailbox, ObservationType, Universe};
-    use tokio::runtime::Runtime;
-
-    use super::*;
-    use crate::actors::combine_partition_ids;
-    use crate::merge_policy::{MergeOperation, StableLogMergePolicy};
-    use crate::new_split_id;
-
-    fn merge_time_range(splits: &[SplitMetadata]) -> Option<RangeInclusive<i64>> {
-        let time_range_start = splits
-            .iter()
-            .flat_map(|split| &split.time_range)
-            .map(|time_range| *time_range.start())
-            .min()?;
-        let time_range_end = splits
-            .iter()
-            .flat_map(|split| &split.time_range)
-            .map(|time_range| *time_range.end())
-            .max()?;
-        Some(time_range_start..=time_range_end)
-    }
-
-    fn merge_tags(splits: &[SplitMetadata]) -> BTreeSet<String> {
-        splits
-            .iter()
-            .flat_map(|split| split.tags.iter().cloned())
-            .collect()
-    }
-
-    fn fake_merge(splits: &[SplitMetadata]) -> SplitMetadata {
-        assert!(!splits.is_empty(), "Split list should not be empty.");
-        let num_docs = splits.iter().map(|split| split.num_docs).sum();
-        let size_in_bytes = splits
-            .iter()
-            .map(|split| split.uncompressed_docs_size_in_bytes)
-            .sum();
-        let time_range = merge_time_range(splits);
-        let merged_split_id = new_split_id();
-        let tags = merge_tags(splits);
-        SplitMetadata {
-            split_id: merged_split_id,
-            partition_id: combine_partition_ids(splits),
-            num_docs,
-            uncompressed_docs_size_in_bytes: size_in_bytes,
-            time_range,
-            create_timestamp: 0,
-            tags,
-            footer_offsets: 0..100,
-            ..Default::default()
-        }
-    }
-
-    fn apply_merge(
-        split_index: &mut HashMap<String, SplitMetadata>,
-        merge_op: &MergeOperation,
-    ) -> Vec<SplitMetadata> {
-        for split in merge_op.splits_as_slice() {
-            assert!(split_index.remove(split.split_id()).is_some());
-        }
-        let splits = vec![fake_merge(merge_op.splits_as_slice())];
-        for split in splits.iter() {
-            split_index.insert(split.split_id().to_string(), split.clone());
-        }
-        splits
-    }
-
-    async fn test_aux_simulate_merge_planner<Pred: Fn(&[SplitMetadata]) -> bool>(
-        merge_policy: Arc<dyn MergePolicy>,
-        incoming_splits: Vec<SplitMetadata>,
-        predicate: Pred,
-    ) -> anyhow::Result<()> {
-        let (merge_op_mailbox, merge_op_inbox) = create_test_mailbox::<MergeSplitDownloader>();
-        let pipeline_id = IndexingPipelineId {
-            index_id: "test-index".to_string(),
-            source_id: "test-source".to_string(),
-            node_id: "test-node".to_string(),
-            pipeline_ord: 0,
-        };
-        let merge_planner =
-            MergePlanner::new(pipeline_id, Vec::new(), merge_policy, merge_op_mailbox);
-        let universe = Universe::new();
-        let mut split_index: HashMap<String, SplitMetadata> = HashMap::default();
-        let (merge_planner_mailbox, merge_planner_handler) =
-            universe.spawn_builder().spawn(merge_planner);
-        for split in incoming_splits {
-            split_index.insert(split.split_id().to_string(), split.clone());
-            merge_planner_mailbox
-                .send_message(NewSplits {
-                    new_splits: vec![split],
-                })
-                .await?;
-            loop {
-                let obs = merge_planner_handler.process_pending_and_observe().await;
-                assert_eq!(obs.obs_type, ObservationType::Alive);
-                let merge_ops: Vec<MergeOperation> = merge_op_inbox
-                    .drain_for_test()
-                    .into_iter()
-                    .flat_map(|op| op.downcast::<MergeOperation>())
-                    .map(|op| *op)
-                    .collect();
-                if merge_ops.is_empty() {
-                    break;
-                }
-                for merge_op in merge_ops {
-                    let splits = apply_merge(&mut split_index, &merge_op);
-                    merge_planner_mailbox
-                        .send_message(NewSplits { new_splits: splits })
-                        .await?;
-                }
-            }
-            let split_metadatas: Vec<SplitMetadata> = split_index.values().cloned().collect();
-            assert!(predicate(&split_metadatas));
-        }
-        Ok(())
-    }
-
-    /// Mock split meta helper.
-    fn mock_split_meta_from_num_docs(
-        time_range: RangeInclusive<i64>,
-        num_docs: u64,
-    ) -> SplitMetadata {
-        SplitMetadata {
-            split_id: crate::new_split_id(),
-            partition_id: 3u64,
-            num_docs: num_docs as usize,
-            uncompressed_docs_size_in_bytes: 256u64 * num_docs,
-            time_range: Some(time_range),
-            create_timestamp: 0,
-            tags: BTreeSet::from_iter(vec!["tenant_id:1".to_string(), "tenant_id:2".to_string()]),
-            footer_offsets: 0..100,
-            ..Default::default()
-        }
-    }
-
-    async fn aux_test_simulate_merge_planner_num_docs<TPred: Fn(&[SplitMetadata]) -> bool>(
-        merge_policy: Arc<dyn MergePolicy>,
-        batch_num_docs: &[usize],
-        predicate: TPred,
-    ) -> anyhow::Result<()> {
-        let split_metadatas: Vec<SplitMetadata> = batch_num_docs
-            .iter()
-            .cloned()
-            .enumerate()
-            .map(|(split_ord, num_docs)| {
-                let time_first = split_ord as i64 * 1_000;
-                let time_last = time_first + 999;
-                let time_range = time_first..=time_last;
-                mock_split_meta_from_num_docs(time_range, num_docs as u64)
-            })
-            .collect();
-        test_aux_simulate_merge_planner(merge_policy, split_metadatas, predicate).await?;
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_simulate_merge_planner_constant_case() -> anyhow::Result<()> {
-        let merge_policy = StableLogMergePolicy::default();
-        aux_test_simulate_merge_planner_num_docs(
-            Arc::new(merge_policy.clone()),
-            &vec![10_000; 100_000],
-            |splits| {
-                let num_docs = splits.iter().map(|split| split.num_docs as u64).sum();
-                splits.len() <= merge_policy.max_num_splits_ideal_case(num_docs)
-            },
-        )
-        .await?;
-        Ok(())
-    }
-
-    use proptest::prelude::*;
-
-    fn proptest_config() -> ProptestConfig {
-        let mut proptest_config = ProptestConfig::with_cases(20);
-        proptest_config.max_shrink_iters = 600;
-        proptest_config
-    }
-
-    proptest! {
-        #![proptest_config(proptest_config())]
-        #[test]
-        fn test_proptest_simulate_stable_log_merge_planner_adversarial(batch_num_docs in proptest::collection::vec(select(&[11, 1_990, 10_000, 50_000, 310_000][..]), 1..1_000)) {
-            let merge_policy = StableLogMergePolicy::default();
-            let rt = Runtime::new().unwrap();
-            rt.block_on(
-            aux_test_simulate_merge_planner_num_docs(
-                Arc::new(merge_policy.clone()),
-                &batch_num_docs,
-                |splits| {
-                    let num_docs = splits.iter().map(|split| split.num_docs as u64).sum();
-                    splits.len() <= merge_policy.max_num_splits_worst_case(num_docs)
-                },
-            )).unwrap();
-        }
-    }
-
-    #[tokio::test]
-    async fn test_simulate_stable_log_merge_planner_ideal_case() -> anyhow::Result<()> {
-        let merge_policy = StableLogMergePolicy::default();
-        aux_test_simulate_merge_planner_num_docs(
-            Arc::new(merge_policy.clone()),
-            &vec![10_000; 1_000],
-            |splits| {
-                let num_docs = splits.iter().map(|split| split.num_docs as u64).sum();
-                splits.len() <= merge_policy.max_num_splits_ideal_case(num_docs)
-            },
-        )
-        .await?;
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_simulate_stable_log_merge_planner_bug() -> anyhow::Result<()> {
-        let merge_policy = StableLogMergePolicy::default();
-        let vals = &[11, 11, 11, 11, 11, 11, 310000, 11, 11, 11, 11, 11, 11, 11];
-        aux_test_simulate_merge_planner_num_docs(
-            Arc::new(merge_policy.clone()),
-            &vals[..],
-            |splits| {
-                let num_docs = splits.iter().map(|split| split.num_docs as u64).sum();
-                splits.len() <= merge_policy.max_num_splits_worst_case(num_docs)
-            },
-        )
-        .await?;
-        Ok(())
-    }
-}
