+++ quickwit/quickwit-indexing/src/actors/mod.rs
@@ -44,7 +44,7 @@ pub use self::indexer::{Indexer, IndexerCounters};
 pub use self::ingest_api_garbage_collector::{
     IngestApiGarbageCollector, IngestApiGarbageCollectorCounters,
 };
-pub use self::merge_executor::{combine_partition_ids, MergeExecutor};
+pub use self::merge_executor::{combine_partition_ids, merge_split_attrs, MergeExecutor};
 pub use self::merge_planner::MergePlanner;
 pub use self::merge_split_downloader::MergeSplitDownloader;
 pub use self::packager::Packager;
