+++ quickwit/quickwit-indexing/src/models/mod.rs
@@ -50,7 +50,7 @@ pub use publish_lock::{NewPublishLock, PublishLock};
 pub use publisher_message::SplitUpdate;
 pub use raw_doc_batch::RawDocBatch;
 pub use scratch_directory::ScratchDirectory;
-pub use split_attrs::SplitAttrs;
+pub use split_attrs::{create_split_metadata, SplitAttrs};
 
 #[derive(Clone, Copy, Debug)]
 pub struct Observe;
